create new file for utility functions
